App Engine/requirements.txt
Flask==2.3.3
google-cloud-pubsub
gunicorn